+++ RapidPay/src/Models/BalanceResult.cs
@@ -4,7 +4,7 @@ public class BalanceResult
 {
     public required string CardNumber { get; set; }
 
-    public float Balance { get; set; }
+    public decimal Balance { get; set; }
 
-    public float? CreditLimit { get; set; }
+    public decimal? CreditLimit { get; set; }
 }
