+++ RapidPay/src/Models/UfeRate.cs
@@ -4,7 +4,7 @@ public class UfeRate
 {
     public DateTime TimeStamp { get; set; }
 
-    public float Rate { get; set; }
+    public decimal Rate { get; set; }
 
-    public float Amout { get; set; }
+    public decimal Amout { get; set; }
 }
